media-source/URL-createObjectURL-null.html

<!doctype html>
<html>
<head>
  <meta charset='utf-8'>
  <title>URL.createObjectURL(null)</title>
  <script src="/resources/testharness.js"></script>
  <script src="/resources/testharnessreport.js"></script>
</head>
<body>
<div id="log"></div>
<script>
test(function() {
    var url = window.URL.createObjectURL(null);
    assert_equals(url, null);
}, "URL.createObjectURL(null)");
</script>
</body>
</html>
